This is code that based on the Viewing frustrum of a camera and importance of objects, optimizes a viewpoint.
The view point is based on collision between objects in the scene to the
depth of field of a camera.  The system will try and optimize to a
solution such that the important objects go to the center of the screen
and stay in the depth of field of the camera.